pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Voice based RAG bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "build==1.3.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "chromadb==0.4.18",
    "click==8.3.0",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "executing==2.2.1",
    "faiss-cpu==1.12.0",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "google-auth>=2.30.0",
    "google-cloud-aiplatform>=1.62.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.26.2",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==7.1.0",
    "ipython==9.7.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "kubernetes==34.1.0",
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-core==0.3.19",
    "langsmith==0.1.143",
    "langchain-text-splitters==0.3.2",
    "markdown-it-py==4.0.0",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.19.2",
    "openai-whisper",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "platformdirs==4.5.0",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.0",
    "psutil==7.1.3",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pydantic==2.9.2",
    "pygments==2.19.2",
    "pymupdf==1.24.14",
    "pypdf==5.1.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pyttsx3",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "sentence-transformers==3.3.1",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy>=1.4,<2.0.36",
    "stack-data==0.6.3",
    "streamlit",
    "sympy==1.14.0",
    "tenacity<9.0.0",
    "threadpoolctl==3.6.0",
    "tiktoken",
    "tokenizers==0.19.1",
    "torch==2.2.2",
    "torchaudio==2.2.2",
    "torchvision==0.17.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.44.2",
    "typer==0.20.0",
    "typer-slim==0.20.0",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
    "pydantic-settings==2.11.0",
    "google-cloud-aiplatform>=1.62.0",
    "google-api-core>=2.20.0",
    "google-cloud-core==2.3.3",
    "langchain-groq==0.2.1",
    "unstructured[xlsx]==0.16.9",
    "openpyxl==3.1.5",
    "flask>=3.1.2",
]
